requirements.txt
# Core audio processing
librosa>=0.10.1
soundfile>=0.12.1
pydub>=0.25.1
numpy>=1.24.0
scipy>=1.11.0

# Video and media handling
moviepy>=1.0.3
yt-dlp>=2024.8.1
ffmpeg-python>=0.2.0

# Machine learning and AI
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.35.0
pyannote.audio>=3.1.0
speechbrain>=0.5.16
openai-whisper>=20231117

# Acoustic analysis
praat-parselmouth>=0.4.3
python-speech-features>=0.6

# Data processing and utilities
pandas>=2.0.0
tqdm>=4.66.0
python-dotenv>=1.0.0

# CLI and configuration
click>=8.1.7
pyyaml>=6.0.1
colorama>=0.4.6

# Visualization (optional, for analysis)
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development tools
black>=23.0.0
ruff>=0.1.0
mypy>=1.5.0

# Additional for Phase 2
pandas>=2.0.0
